homework_bot/requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.1
attrs==22.2.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.6
flake8==3.9.2
flake8-docstrings==1.6.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.6.1
packaging==23.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.7.0
pydocstyle==6.3.0
pyflakes==2.3.1
pytest==6.2.5
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
tornado==6.2
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
